compiler fix

+++ cpp/src/barretenberg/stdlib/primitives/field/field.test.cpp
@@ -995,7 +995,7 @@ template <typename Composer> class stdlib_field : public testing::Test {
                 break;
             }
             case 2: {
-                a_val += (a_val == 0) ? 1 : 0;
+                a_val += static_cast<uint8_t>((a_val == 0) ? static_cast<uint8_t>(1) : static_cast<uint8_t>(0));
                 b_val = a_val - 1;
                 break;
             }
